Split cmd/crane implementation to pkg/crane
Rewrites each "init()" to be "NewCmdFoo". This pattern was cribbed from
the kubectl command in kubernetes/kubernetes.

I'm doing this so I can reuse most of these for cmd/gcr and selectively
replace bits of it with gcr-specific stuff.

I can forsee us cleaning up the signatures of these crane cmd
implementations to be more friendly and then exposing them so that
folks don't have to rewrite these common functions over and over again.
