+++ cmd/crane/BUILD.bazel
@@ -4,28 +4,12 @@ load("@io_bazel_rules_docker//go:image.bzl", "go_image")
 go_library(
     name = "go_default_library",
     srcs = [
-        "append.go",
-        "config.go",
-        "copy.go",
-        "delete.go",
-        "digest.go",
-        "get.go",
-        "list.go",
         "main.go",
-        "manifest.go",
-        "pull.go",
-        "push.go",
-        "rebase.go",
     ],
     importpath = "github.com/google/go-containerregistry/cmd/crane",
     visibility = ["//visibility:private"],
     deps = [
-        "//authn:go_default_library",
-        "//name:go_default_library",
-        "//v1:go_default_library",
-        "//v1/mutate:go_default_library",
-        "//v1/remote:go_default_library",
-        "//v1/tarball:go_default_library",
+        "//pkg/crane:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
     ],
 )
