+++ pkg/crane/BUILD.bazel
@@ -1,0 +1,29 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "append.go",
+        "config.go",
+        "copy.go",
+        "delete.go",
+        "digest.go",
+        "get.go",
+        "list.go",
+        "manifest.go",
+        "pull.go",
+        "push.go",
+        "rebase.go",
+    ],
+    importpath = "github.com/google/go-containerregistry/pkg/crane",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//authn:go_default_library",
+        "//name:go_default_library",
+        "//v1:go_default_library",
+        "//v1/mutate:go_default_library",
+        "//v1/remote:go_default_library",
+        "//v1/tarball:go_default_library",
+        "//vendor/github.com/spf13/cobra:go_default_library",
+    ],
+)
