+++ pkg/crane/digest.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package crane
 
 import (
 	"fmt"
@@ -21,24 +21,24 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func init() {
-	rootCmd.AddCommand(&cobra.Command{
-		Use:   "config",
-		Short: "Get the config of an image",
+func NewCmdDigest() *cobra.Command {
+	return &cobra.Command{
+		Use:   "digest",
+		Short: "Get the digest of an image",
 		Args:  cobra.ExactArgs(1),
-		Run:   config,
-	})
+		Run:   digest,
+	}
 }
 
-func config(_ *cobra.Command, args []string) {
+func digest(_ *cobra.Command, args []string) {
 	ref := args[0]
 	i, _, err := getImage(ref)
 	if err != nil {
 		log.Fatalln(err)
 	}
-	config, err := i.RawConfigFile()
+	digest, err := i.Digest()
 	if err != nil {
 		log.Fatalln(err)
 	}
-	fmt.Print(string(config))
+	fmt.Print(digest.String())
 }
